Initial SME Support - SE-122  (#248)
This PR adds functionality into SimEng to support the AArch64 SME extension. 6 new SME instructions have been implemented, along with the SVCR Streaming-SVE-Mode context switching functionality. 

Additionally, the default LLVM version has been updated to 14.0.5 in order to support the SME regression tests.
The AArch64 Instruction class now has a copy constructor, optimising the use of cached instructions.
Updated Jenkin's pipeline scripts to work with LLVM 14
Changes the Capstone usage to fix an existing memory leak, as mentioned in Issue #261.

+++ src/include/simeng/pipeline/RenameUnit.hh
@@ -19,7 +19,7 @@ class RenameUnit {
   RenameUnit(PipelineBuffer<std::shared_ptr<Instruction>>& input,
              PipelineBuffer<std::shared_ptr<Instruction>>& output,
              ReorderBuffer& rob, RegisterAliasTable& rat, LoadStoreQueue& lsq,
-             uint8_t registerTypes);
+             uint16_t registerTypes);
 
   /** Ticks this unit. Renames registers of instructions, and allocates ROB
    * space. */
@@ -59,7 +59,7 @@ class RenameUnit {
 
   /** A table recording the numbers of free physical registers for each register
    * file. */
-  std::vector<uint8_t> freeRegistersAvailable_;
+  std::vector<uint16_t> freeRegistersAvailable_;
 
   /** The number of cycles stalled due to inability to allocate enough
    * destination registers. */
